backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
python-dotenv>=1.0.0
